Create constellation code
In this example, we use the @google-cloud/vision package to interact with the Google Cloud Vision API. Make sure to install the package by running npm install @google-cloud/vision before executing the code. Note that you'll need to set up the Google Cloud Vision API and provide the necessary credentials for authentication.